+++ rust/metatensor/src/data/array_ref.rs
@@ -223,11 +223,11 @@ impl<'a> ArrayRef<'a> {
     /// Copy the data in this array, if supported by the underlying data.
     ///
     /// This corresponds to `mts_array_t.copy`, but with a more convenient API.
-    pub fn copy(&self) -> Result<MtsArray, Error> {
+    pub fn copy(&self, device: DLDevice) -> Result<MtsArray, Error> {
         let function = self.array.copy.expect("mts_array_t.copy function is NULL");
         let mut new_array = mts_array_t::null();
         unsafe {
-            check_status(function(self.array.ptr, &mut new_array))?;
+            check_status(function(self.array.ptr, device, &mut new_array))?;
         }
 
         return Ok(MtsArray::from_raw(new_array));
@@ -521,11 +521,11 @@ impl<'a> ArrayRefMut<'a> {
     /// Copy the data in this array, if supported by the underlying data.
     ///
     /// This corresponds to `mts_array_t.copy`, but with a more convenient API.
-    pub fn copy(&self) -> Result<MtsArray, Error> {
+    pub fn copy(&self, device: DLDevice) -> Result<MtsArray, Error> {
         let function = self.array.copy.expect("mts_array_t.copy function is NULL");
         let mut new_array = mts_array_t::null();
         unsafe {
-            check_status(function(self.array.ptr, &mut new_array))?;
+            check_status(function(self.array.ptr, device, &mut new_array))?;
         }
 
         return Ok(MtsArray::from_raw(new_array));